test(musa): forbid default-on process gates
